server/requirements.txt
absl-py==1.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
astor==0.8.1
beautifulsoup4==4.12.3
blurhash==1.1.4
boost==0.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
crcmod==1.7
cryptography==40.0.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
deepface==0.0.86
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.3
fer==22.5.1
ffmpeg==1.4
filelock==3.4.1
fire==0.6.0
Flask==2.0.3
gast==0.6.0
gdown==4.7.3
greenlet==2.0.2
grpcio==1.48.2
grpcio-tools==1.48.2
gunicorn==21.2.0
h5py==2.7.0
idna==3.7
imageio==2.15.0
imageio-ffmpeg==0.5.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
Mastodon.py==1.8.1
matplotlib==3.3.4
moviepy==1.0.3
mtcnn==0.1.1
opencv-contrib-python==3.4.4.19
opencv-python==4.6.0.66
oss2==2.18.6
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
proglog==0.1.10
protobuf==3.19.6
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
retina-face==0.0.17
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.52
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.19
Werkzeug==2.0.3
zipp==3.6.0
numpy
dlib
